Use BigInteger class

+++ scray-hdfs-writer/src/main/scala/scray/hdfs/index/format/Writer.scala
@@ -17,13 +17,14 @@ package scray.hdfs.index.format
 
 import scray.hdfs.index.format.sequence.types.Blob
 import java.io.InputStream
+import java.math.BigInteger
 
 trait Writer { 
   var varIsClosed = false
   
   def insert(id: String, updateTime: Long, data: Array[Byte]): Long
   def insert(id: String, updateTime: Long, data: InputStream, blobSplitSize: Int = 5 * 1024 * 1024): Long
-  def insert(id: String, updateTime: Long, data: InputStream, dataSize: Int, blobSplitSize: Int): Long
+  def insert(id: String, updateTime: Long, data: InputStream, dataSize: BigInteger, blobSplitSize: Int): Long
   def insert(idBlob: Tuple2[String, Blob]): Unit
   def getBytesWritten: Long
   def getNumberOfInserts: Int
